01-agentic-rag-assistant/requirements.txt
gradio==6.3.0

langchain>=0.1.0
langchain-core>=0.1.7
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.2
langchain-groq>=1.1.1

sentence-transformers>=3.1.1
faiss-cpu>=1.13.2
rank-bm25>=0.2.2

openai>=1.12.0
tiktoken>=0.5.2

pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
numpy>=2.1.0
pandas>=2.0.0

